feat(checkbox): extract common logic in a directive
