+++ projects/igniteui-angular/src/lib/radio/radio.component.ts
@@ -10,8 +10,7 @@ import {
 import { ControlValueAccessor } from '@angular/forms';
 import { EditorProvider, EDITOR_PROVIDER } from '../core/edit-provider';
 import { IgxRippleDirective } from '../directives/ripple/ripple.directive';
-import { IgxCheckboxComponent } from '../checkbox/checkbox.component';
-
+import { CheckboxBaseDirective } from '../checkbox/checkbox-base.directive';
 
 /**
  * **Ignite UI for Angular Radio Button** -
@@ -36,7 +35,9 @@ import { IgxCheckboxComponent } from '../checkbox/checkbox.component';
     templateUrl: 'radio.component.html',
     imports: [IgxRippleDirective]
 })
-export class IgxRadioComponent extends IgxCheckboxComponent implements AfterViewInit, ControlValueAccessor, EditorProvider {
+export class IgxRadioComponent
+    extends CheckboxBaseDirective
+    implements AfterViewInit, ControlValueAccessor, EditorProvider {
     /** @hidden @internal */
     public blurRadio = new EventEmitter();
 
@@ -149,7 +150,11 @@ export class IgxRadioComponent extends IgxCheckboxComponent implements AfterView
     public select() {
         if (!this.checked) {
             this.checked = true;
-            this.change.emit({ value: this.value, owner: this, checked: this.checked });
+            this.change.emit({
+                value: this.value,
+                owner: this,
+                checked: this.checked,
+            });
             this._onChangeCallback(this.value);
         }
     }
