+++ projects/igniteui-angular/src/lib/switch/switch.component.ts
@@ -8,7 +8,7 @@ import {
 import { ControlValueAccessor } from '@angular/forms';
 import { IgxRippleDirective } from '../directives/ripple/ripple.directive';
 import { EditorProvider, EDITOR_PROVIDER } from '../core/edit-provider';
-import { IgxCheckboxComponent } from '../checkbox/checkbox.component';
+import { CheckboxBaseDirective } from '../checkbox/checkbox-base.directive';
 
 /**
  *
@@ -36,13 +36,15 @@ import { IgxCheckboxComponent } from '../checkbox/checkbox.component';
     providers: [{
             provide: EDITOR_PROVIDER,
             useExisting: IgxSwitchComponent,
-            multi: true
-        }],
+        multi: true
+    }],
     selector: 'igx-switch',
     templateUrl: 'switch.component.html',
     imports: [IgxRippleDirective]
 })
-export class IgxSwitchComponent extends IgxCheckboxComponent implements ControlValueAccessor, EditorProvider, AfterViewInit {
+export class IgxSwitchComponent
+    extends CheckboxBaseDirective
+    implements ControlValueAccessor, EditorProvider, AfterViewInit {
     /**
      * Returns the class of the switch component.
      *
